module/requirements.txt
pytest
google-cloud-bigquery
Jinja2